tg-8634 [BE] Modificar endpoint summary-list para que incluya institucion y tipo de cierre en contrarreferencia

+++ back-end/reference-counter-ref/src/main/java/ar/lamansys/refcounterref/domain/counterreference/CounterReferenceBo.java
@@ -41,5 +41,5 @@ public class CounterReferenceBo {
 
     private List<Integer> fileIds;
 
-	private Integer closureTypeId;
+	private Short closureTypeId;
 }
